+++ test/js/esm/ssr.js
@@ -10,7 +10,7 @@
 export function render(page) {
   return {
     head: [
-      `<title data-inertia>New title from ESM</title>`,
+      `<title data-inertia="">New title from ESM</title>`,
       `<meta name="description" content="Head stuff" />`,
     ],
     body: `<div id="ssr">${page.props.content || ""}</div>`,